% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootMSE.meteDist.R
\name{bootMSE.meteDist}
\alias{bootMSE.meteDist}
\title{Bootstrap the mean squared error of a `meteR::meteDist` object}
\usage{
bootMSE.meteDist(x, B = 999, type = "rank", relative = TRUE,
  log = FALSE)
}
\arguments{
\item{x}{a `meteDist` object}

\item{B}{number of bootstrap replicates}

\item{type}{should the error be calculated on the rank distribution of cumulative distribution}

\item{relative}{logical, should the error be calculated on a relative scale}

\item{log}{logical, should the error be calculated on a log scale}
}
\description{
Produce bootstrap sampling distribution of the MSE of a `meteDist` object
}
\details{
Currently only SAD and IPD classes are supported
}
\examples{
x <- rpois(100, 1) + 1 # this won't look very METEish
esf <- meteESF(1:length(x), x)
bootSAD <- bootMSE.meteDist(sad(esf))

}
\author{
A.J. Rominger <ajrominger@gmail.com>
}
