% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{grun}
\alias{grun}
\title{Arthropod community abundance and body size data from across the Hawaiian archipelago}
\format{A data frame with 15997 rows and 19 columns:
\describe{
  \item{SpecimenCode}{specimen ID}
  \item{XID}{foo}
  \item{CollectionCode}{collection ID}
  \item{Site}{collection site}
  \item{Tree}{collection tree ID}
  \item{Tray}{collection tray ID}
  \item{DateCollected}{collection date}
  \item{Class}{taxonomic class}
  \item{Order}{taxonomic class}
  \item{Family}{taxonomic class}
  \item{SpeciesCode}{unique species ID}
  \item{Genus}{taxonomic genus}
  \item{species}{species epithet}
  \item{trophic}{trophic level: `D` = detritivore, `H` = herbivore, `P` = predator, `T` = transient, `U` = unknown}
  \item{Origin}{geographic origin}
  \item{Stage}{life stage}
  \item{Lengthmm}{length in mm}
  \item{Abundance}{abundance in the sample}
  \item{IND_BIOM}{estimated individual biomass in mg}
}}
\source{
Gruner, D. S. 2007. Geological age, ecosystem development, and local resource constraints on arthropod community structure in the Hawaiian Islands. Biological Journal of the Linnean Society, 90: 551--570.
}
\usage{
grun
}
\description{
A dataset containing the community abundace data for individuals, as well 
as their body mass.
}
\keyword{datasets}
