% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{grunSite}
\alias{grunSite}
\title{Arthropod collection site information}
\format{A data frame with 5 rows and 5 columns:
\describe{
  \item{code}{collection site ID as it appears in `grun`}
  \item{name}{full name of the collection site}
  \item{age}{substrate age of the collection site in millions of years}
  \item{island}{island where the collection site is located}
  \item{island_age}{geologic age of the island in millions of years}
}}
\usage{
grunSite
}
\description{
A dataset with geologic and geographic information on the colleciton sites
}
\keyword{datasets}
