% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.meteDist.R
\name{sample.meteDist}
\alias{sample.meteDist}
\title{Sample from a `meteR::meteDist` object}
\usage{
sample.meteDist(x, size, replace = FALSE)
}
\arguments{
\item{x}{a `meteDist` object}

\item{size}{number of individuals to sample}

\item{replace}{logical, whether sampling should be with or without replacement}
}
\description{
Draw a random sample of individuals from a `meteDist` object, with or 
without replacement.
}
\details{
Currently only SAD and IPD classes are supported
}
\examples{
x <- rpois(100, 1) + 1 # this won't look very METEish
esf <- meteESF(1:length(x), x)
subSAD <- sample.meteDist(sad(esf), size = round(0.5 * sum(x)), replace = TRUE)

}
\author{
A.J. Rominger <ajrominger@gmail.com>
}
